+++ core/src/main/java/de/pottgames/tuningfork/decoder/WavInputStream.java
@@ -1,48 +1,47 @@
 /**
  * Copyright 2022 Matthias Finke
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the
- * License at
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 
 package de.pottgames.tuningfork.decoder;
 
-import java.io.IOException;
-import java.io.InputStream;
-
 import com.badlogic.gdx.files.FileHandle;
 import com.badlogic.gdx.utils.StreamUtils;
-
 import de.pottgames.tuningfork.Audio;
 import de.pottgames.tuningfork.PcmFormat;
 import de.pottgames.tuningfork.PcmFormat.PcmDataType;
 import de.pottgames.tuningfork.TuningForkRuntimeException;
 import de.pottgames.tuningfork.logger.TuningForkLogger;
 
+import java.io.IOException;
+import java.io.InputStream;
+
 /**
  * An {@link AudioStream} implementation to read wav files.
  *
  * @author Matthias
- *
  */
 public class WavInputStream implements AudioStream {
     private final InputStream      stream;
-    private WavDecoder             decoder;
+    private       WavDecoder       decoder;
     private final TuningForkLogger logger;
     private final FileHandle       file;
     private final float            duration;
-    private boolean                closed = false;
+    private       boolean          closed = false;
 
 
     /**
      * Initializes a {@link WavInputStream} from a {@link FileHandle}.
      *
-     * @param file
+     * @param file the file handle
      */
     public WavInputStream(FileHandle file) {
         this(file, true);
@@ -52,7 +51,7 @@ public class WavInputStream implements AudioStream {
     /**
      * Initializes a {@link WavInputStream} from a {@link FileHandle}.
      *
-     * @param file
+     * @param file         the file handle
      * @param forStreaming true if this will be used for streaming
      */
     public WavInputStream(FileHandle file, boolean forStreaming) {
@@ -65,10 +64,10 @@ public class WavInputStream implements AudioStream {
 
 
     /**
-     * Initializes a {@link WavInputStream} from an {@link InputStream}. This stream does not support the reset function. Use
-     * {@link #WavInputStream(FileHandle)} instead to get the full functionality.
+     * Initializes a {@link WavInputStream} from an {@link InputStream}. This stream does not support the reset
+     * function. Use {@link #WavInputStream(FileHandle)} instead to get the full functionality.
      *
-     * @param stream
+     * @param stream the input stream
      */
     public WavInputStream(InputStream stream) {
         this(stream, true);
@@ -76,10 +75,10 @@ public class WavInputStream implements AudioStream {
 
 
     /**
-     * Initializes a {@link WavInputStream} from an {@link InputStream}. This stream does not support the reset function. Use
-     * {@link #WavInputStream(FileHandle)} instead to get the full functionality.
+     * Initializes a {@link WavInputStream} from an {@link InputStream}. This stream does not support the reset
+     * function. Use {@link #WavInputStream(FileHandle)} instead to get the full functionality.
      *
-     * @param stream
+     * @param stream       the input stream
      * @param forStreaming true if this will be used for streaming
      */
     public WavInputStream(InputStream stream, boolean forStreaming) {
@@ -107,7 +106,8 @@ public class WavInputStream implements AudioStream {
             this.throwRuntimeError("Unsupported wav file format");
         }
         this.decoder.setup(this.stream, bytesRemaining);
-        if (PcmFormat.determineFormat(this.decoder.outputChannels(), this.decoder.outputBitsPerSample(), this.decoder.outputPcmDataType()) == null) {
+        if (PcmFormat.determineFormat(this.decoder.outputChannels(), this.decoder.outputBitsPerSample(),
+                                      this.decoder.outputPcmDataType()) == null) {
             this.throwRuntimeError("Unsupported format found in wav file");
         }
     }
@@ -116,7 +116,8 @@ public class WavInputStream implements AudioStream {
     private void readRiffChunk() {
         try {
             // RIFF LITERAL
-            final boolean riff = this.stream.read() == 'R' && this.stream.read() == 'I' && this.stream.read() == 'F' && this.stream.read() == 'F';
+            final boolean riff = this.stream.read() == 'R' && this.stream.read() == 'I' && this.stream.read() == 'F' &&
+                                 this.stream.read() == 'F';
             if (!riff) {
                 this.throwRuntimeError("Not a valid wav file, RIFF header missing");
             }
@@ -128,7 +129,8 @@ public class WavInputStream implements AudioStream {
             }
 
             // WAVE LITERAL
-            final boolean wave = this.stream.read() == 'W' && this.stream.read() == 'A' && this.stream.read() == 'V' && this.stream.read() == 'E';
+            final boolean wave = this.stream.read() == 'W' && this.stream.read() == 'A' && this.stream.read() == 'V' &&
+                                 this.stream.read() == 'E';
             if (!wave) {
                 this.throwRuntimeError("Not a valid wav file, WAVE literal missing");
             }
@@ -141,7 +143,8 @@ public class WavInputStream implements AudioStream {
     private WavFmtChunk readFmtChunk() {
         try {
             // FMT LITERAL
-            final boolean fmt = this.stream.read() == 'f' && this.stream.read() == 'm' && this.stream.read() == 't' && this.stream.read() == ' ';
+            final boolean fmt = this.stream.read() == 'f' && this.stream.read() == 'm' && this.stream.read() == 't' &&
+                                this.stream.read() == ' ';
             if (!fmt) {
                 this.throwRuntimeError("Not a valid wav file, FMT  header missing");
             }
@@ -174,7 +177,7 @@ public class WavInputStream implements AudioStream {
         if (byte1 < 0 || byte2 < 0 || byte3 < 0 || byte4 < 0) {
             return -1L;
         }
-        return byte1 | byte2 << 8L | byte3 << 16L | byte4 << 24L;
+        return byte1 | byte2 << 8L | byte3 << 16L | (long) byte4 << 24L;
     }
